add proper error model for non-existent paths
