+++ shared/fs/index.stories.js
@@ -5,7 +5,7 @@ import footerStories from './footer/index.stories'
 import sendLinkToChatStories from './send-link-to-chat/index.stories'
 import destinationPickerStories from './destination-picker/index.stories'
 import folderStories from './folder/index.stories'
-import oopsNoAccessStories from './oops-no-access/index.stories'
+import oopsStories from './oops/index.stories'
 
 export default () =>
   [
@@ -15,5 +15,5 @@ export default () =>
     footerStories,
     destinationPickerStories,
     sendLinkToChatStories,
-    oopsNoAccessStories,
+    oopsStories,
   ].forEach(l => l())
